Apply historical population based on population type capacities for each holding

+++ src/game/game.h
@@ -116,7 +116,7 @@ public:
 	void apply_sites();
 	void apply_site_buildings(const site *site);
 	void apply_population_history();
-	int64_t apply_historical_population_group_to_site(const population_group_key &group_key, const int population, const site *site);
+	int64_t apply_historical_population_group_to_site(const population_group_key &group_key, const int64_t population, const site *site);
 
 	QCoro::Task<void> on_setup_finished();
